+++ libcontainer/cgroups/devices/config/device.go
@@ -1,4 +1,4 @@
-package devices
+package config
 
 import (
 	"fmt"
